Fix crossOrigin for resources on same server
Since 814607053688eed3f44a26e6116fad8ce42f2197 the module uses `url` as a dependency, which cause an issue since `url` returns `null` when the `port` is not defined. Whereas `window.location.port` returns an empty string `''`.
The module used `anonymous` even if the resources were on the same server.