Moved dequeue to after push to _resourcesParsing (#92)
* Moved dequeue to after push to `_resourcesParsing`

A situation can arise when there's 2 resources being loaded with the second one
being a spritesheet. Spritesheets require an `after` middleware to add extra
files to be loaded (the spritesheet image itself).

If the call to `dequeue` happens on the spritesheet resource and the `_onLoad`
of the previous resource is on the queue, then the `_onLoad` is resumed. The
`after` middleware will be run the first resource and when it is spliced out of
`_resourcesParsing` in the `async.eachSeries` completion callback the
spritesheet resource will not be in `_resourcesParsing`. `_resourcesParsing` is
now empty and `_onComplete` is called. The `after` middleware is not run before
on the spritesheet and thus it hasn't loaded its associated texture.

If you have code relying on the texture to be loaded when `onComplete` is called
then it will not be loaded at that point.

The fix is to move the push to `_resourcesParsing` to before the `dequeue` in
order to ensure its present when any previously queued `_onLoad` calls are
executed

* Added test case for middleware race condition

* Fixed linter errors